Refine Aptos e2e test lane configs
