+++ deployment/ccip/changeset/testhelpers/test_helpers_solana_v0_1_0.go
@@ -1160,131 +1160,6 @@ func AddEVMDestChangesets(e *DeployedEnv, to, from uint64, isTestRouter bool) []
 	return evmDstChangesets
 }
 
-func AddLaneAptosChangesets(t *testing.T, srcChainSelector, destChainSelector uint64, gasPrices map[uint64]*big.Int, tokenPrices map[aptos.AccountAddress]*big.Int) []commoncs.ConfiguredChangeSet {
-	srcFamily, err := chainsel.GetSelectorFamily(srcChainSelector)
-	require.NoError(t, err)
-	destFamily, err := chainsel.GetSelectorFamily(destChainSelector)
-	require.NoError(t, err)
-
-	if srcFamily != chainsel.FamilyAptos &&
-		destFamily != chainsel.FamilyAptos {
-		t.Fatalf("At least one of the provided source/destination chains has to be Aptos. srcFamily: %v destFamily: %v", srcFamily, destFamily)
-	}
-
-	var src, dest config.ChainDefinition
-
-	switch srcFamily {
-	case chainsel.FamilyEVM:
-		src = config.EVMChainDefinition{
-			ChainDefinition: v1_6.ChainDefinition{
-				ConnectionConfig: v1_6.ConnectionConfig{
-					RMNVerificationDisabled: true,
-				},
-				Selector: srcChainSelector,
-			},
-		}
-	case chainsel.FamilyAptos:
-		src = config.AptosChainDefinition{
-			TokenPrices: tokenPrices,
-			ConnectionConfig: v1_6.ConnectionConfig{
-				RMNVerificationDisabled: true,
-			},
-			Selector:                      srcChainSelector,
-			AddTokenTransferFeeConfigs:    nil,
-			RemoveTokenTransferFeeConfigs: nil,
-		}
-	default:
-		t.Fatalf("Unsupported source chain family: %v", srcFamily)
-	}
-
-	switch destFamily {
-	case chainsel.FamilyEVM:
-		dest = config.EVMChainDefinition{
-			ChainDefinition: v1_6.ChainDefinition{
-				ConnectionConfig: v1_6.ConnectionConfig{
-					AllowListEnabled: false,
-				},
-				Selector: destChainSelector,
-				GasPrice: gasPrices[destChainSelector],
-				FeeQuoterDestChainConfig: fee_quoter.FeeQuoterDestChainConfig{
-					IsEnabled:                         true,
-					MaxNumberOfTokensPerMsg:           10,
-					MaxDataBytes:                      30_000,
-					MaxPerMsgGasLimit:                 3_000_000,
-					DestGasOverhead:                   ccipevm.DestGasOverhead,
-					DestGasPerPayloadByteBase:         ccipevm.CalldataGasPerByteBase,
-					DestGasPerPayloadByteHigh:         ccipevm.CalldataGasPerByteHigh,
-					DestGasPerPayloadByteThreshold:    ccipevm.CalldataGasPerByteThreshold,
-					DestDataAvailabilityOverheadGas:   100,
-					DestGasPerDataAvailabilityByte:    16,
-					DestDataAvailabilityMultiplierBps: 1,
-					ChainFamilySelector:               [4]byte{0x28, 0x12, 0xd5, 0x2c},
-					EnforceOutOfOrder:                 false,
-					DefaultTokenFeeUSDCents:           25,
-					DefaultTokenDestGasOverhead:       90_000,
-					DefaultTxGasLimit:                 200_000,
-					GasMultiplierWeiPerEth:            11e8, // TODO what's the scale here ?
-					GasPriceStalenessThreshold:        0,
-					NetworkFeeUSDCents:                10,
-				},
-			},
-			OnRampVersion: []byte{1, 6, 0},
-		}
-	case chainsel.FamilyAptos:
-		dest = config.AptosChainDefinition{
-			ConnectionConfig: v1_6.ConnectionConfig{
-				AllowListEnabled: false,
-			},
-			Selector: destChainSelector,
-			GasPrice: gasPrices[destChainSelector],
-			FeeQuoterDestChainConfig: aptos_fee_quoter.DestChainConfig{
-				IsEnabled:                         true,
-				MaxNumberOfTokensPerMsg:           10,
-				MaxDataBytes:                      30_000,
-				MaxPerMsgGasLimit:                 3_000_000,
-				DestGasOverhead:                   ccipevm.DestGasOverhead,
-				DestGasPerPayloadByteBase:         ccipevm.CalldataGasPerByteBase,
-				DestGasPerPayloadByteHigh:         ccipevm.CalldataGasPerByteHigh,
-				DestGasPerPayloadByteThreshold:    ccipevm.CalldataGasPerByteThreshold,
-				DestDataAvailabilityOverheadGas:   100,
-				DestGasPerDataAvailabilityByte:    16,
-				DestDataAvailabilityMultiplierBps: 1,
-				ChainFamilySelector:               []byte{0xac, 0x77, 0xff, 0xec},
-				EnforceOutOfOrder:                 true,
-				DefaultTokenFeeUsdCents:           25,
-				DefaultTokenDestGasOverhead:       90_000,
-				DefaultTxGasLimit:                 200_000,
-				GasMultiplierWeiPerEth:            11e17,
-				GasPriceStalenessThreshold:        0,
-				NetworkFeeUsdCents:                10,
-			},
-		}
-	default:
-		t.Fatalf("Unsupported dstination chain family: %v", srcFamily)
-	}
-
-	return []commoncs.ConfiguredChangeSet{
-		commoncs.Configure(
-			aptoscs.AddAptosLanes{},
-			config.UpdateAptosLanesConfig{
-				AptosMCMSConfig: &proposalutils.TimelockConfig{
-					MinDelay:     time.Second,
-					MCMSAction:   mcmstypes.TimelockActionSchedule,
-					OverrideRoot: false,
-				},
-				Lanes: []config.LaneConfig{
-					{
-						Source:     src,
-						Dest:       dest,
-						IsDisabled: false,
-					},
-				},
-				TestRouter: false,
-			},
-		),
-	}
-}
-
 // RemoveLane removes a lane between the source and destination chains in the deployed environment.
 func RemoveLane(t *testing.T, e *DeployedEnv, src, dest uint64, isTestRouter bool) {
 	var err error
